v1.53 (2004-12-31)
When the font subdirectory is in the same directory as fpdf.php, it's no longer necessary to define the FPDF_FONTPATH constant.

The array $HTTP_SERVER_VARS is no longer used. It could cause trouble on PHP5-based configurations with the register_long_arrays option disabled.

Fixed a problem related to Type1 font embedding which caused trouble to some PDF processors.

The file name sent to the browser could not contain a space character.

The Cell() method could not print the number 0 (you had to pass the string '0').