v1.70 (2011-06-18)
The MakeFont utility has been completely rewritten and doesn't depend on ttf2pt1 anymore.

Alpha channel is now supported for PNGs.

When inserting an image, it's now possible to specify its resolution.

Default resolution for images was increased from 72 to 96 dpi.

When inserting a GIF image, no temporary file is used anymore if the PHP version is 5.1 or higher.

When output buffering is enabled and the PDF is about to be sent, the buffer is now cleared if it contains only a UTF-8 BOM and/or whitespace (instead of throwing an error).

Symbol and ZapfDingbats fonts now support underline style.

Custom page sizes are now checked to ensure that width is smaller than height.

Standard font files were changed to use the same format as user fonts.

A bug in the embedding of Type1 fonts was fixed.

A bug related to SetDisplayMode() and the current locale was fixed.

A display issue occurring with the Adobe Reader X plug-in was fixed.

An issue related to transparency with some versions of Adobe Reader was fixed.

The Content-Length header was removed because it caused an issue when the HTTP server applies compression.