v1.80 (2015-11-29)
PHP 5.1.0 or higher is now required.

The MakeFont utility now subsets fonts, which can greatly reduce font sizes.

Added ToUnicode CMaps to improve text extraction.

Added a parameter to AddPage() to rotate the page.

Added a parameter to SetY() to indicate whether the x position should be reset or not.

Added a parameter to Output() to specify the encoding of the name, and special characters are now properly encoded. Additionally the order of the first two parameters was reversed to be more logical (however the old order is still supported for compatibility).

The Error() method now throws an exception.

Adding contents before the first AddPage() or after Close() now raises an error.

Outputting text with no font selected now raises an error.